fix: cast Cov to float32 before inverse for AMP compatibility
linalg.inv does not support half precision. With AMP enabled, Cov
would be fp16, causing RuntimeError. Cast to float32 for inversion
then back to original dtype.

+++ model/head/gaussian_head.py
@@ -89,7 +89,7 @@ class GaussianHead(BaseTaskHead):
         R = get_rotation_matrix(rotations) # b, g, 3, 3
         M = torch.matmul(S, R)
         Cov = torch.matmul(M.transpose(-1, -2), M)
-        CovInv = Cov.cpu().inverse().cuda() # b, g, 3, 3
+        CovInv = Cov.float().cpu().inverse().cuda().to(Cov.dtype) # b, g, 3, 3; float() for AMP compatibility
         return means, origi_opa, opacities, scales, CovInv
 
     def get_filtered_lidar(self,lidar):
